requirements.txt
Django==3.1.5
psycopg2-binary==2.8.6
python-dotenv==0.15.0
envparse==0.2.0

redis==3.5.3
django-redis==4.12.1

django-split-settings==1.0.1

wagtail